Added regrid to edit data menu for manipulations
